# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Management System API",
        "description": "RESTful API для системы управления блогом с аутентификацией и системой ролей",
        "contact": {
            "email": "u.azamat.1997@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/blog/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Список всех категорий",
                "operationId": "8c15c4907bf242e869e37154c9eb1468",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Фильтр по активности",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Создание новой категории",
                "operationId": "bcd797fc4179299f1f26c171146b7efe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Технологии"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "texnologii"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание категории"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#3B82F6"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категория создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Категория успешно создана"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/blog/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить категорию по ID",
                "operationId": "bd7bbb1aad8fa5176a30b042c9773e6e",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{category}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Обновить категорию",
                "operationId": "50b29fbace4d4d3acf9b59b1000e1446",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Технологии"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "texnologii"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание категории"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#3B82F6"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно обновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Категория успешно обновлена"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Удалить категорию",
                "operationId": "38eff19b342f7aa1a6019379ccff3b72",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Категория успешно удалена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Технологии"
                    },
                    "slug": {
                        "type": "string",
                        "example": "texnologii"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание категории"
                    },
                    "color": {
                        "type": "string",
                        "example": "#3B82F6"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}